requirements.txt
fastapi
uvicorn
pydantic-ai
google-generativeai
psycopg2-binary
python-dotenv
sentence-transformers
pypdf
pymongo
requests